Split the implementation classes of `MatrixRoom` too
